FIX: MortalStageData [Range] на long убран + GetRandomQiCapacity без int round-trip; IntegrationTests AcceptQi long→long
- MortalStageData: [Range(1,500)] убран с long полей (не работает в Inspector)
- GetRandomQiCapacity: безопасная генерация для long значений (>int.MaxValue через интерполяцию)
- IntegrationTests: long accepted вместо (int)pool.AcceptQi — устранена потеря точности
- Дата редактирования: 2026-04-13 09:25:00 UTC

+++ UnityProject/Assets/Scripts/Tests/IntegrationTests.cs
@@ -504,7 +504,8 @@ namespace CultivationGame.Tests
 
                 // === Act ===
                 int transferAmount = 500;
-                int accepted = (int)pool.AcceptQi(transferAmount, transferRate); // FIX CS0266 (2026-04-11)
+                // Редактировано: 2026-04-13 — long вместо int (AcceptQi возвращает long)
+                long accepted = pool.AcceptQi(transferAmount, transferRate);
 
                 // Тратим Ци у практика
                 if (accepted > 0)
@@ -748,7 +749,8 @@ namespace CultivationGame.Tests
             FormationQiPool pool = new FormationQiPool(1000, 60, 10, 0f); // 0 conductivity
 
             // === Act ===
-            int accepted = (int)pool.AcceptQi(100, 50f); // FIX CS0266 (2026-04-11)
+            // Редактировано: 2026-04-13 — long вместо int (AcceptQi возвращает long)
+            long accepted = pool.AcceptQi(100, 50f);
 
             // === Assert ===
             // Даже с нулевой проводимостью пул должен принять некоторое количество
